Bug 177210  - Stack overflow in StyledTextRenderer

+++ bundles/org.eclipse.swt/Eclipse SWT/wpf/org/eclipse/swt/graphics/TextLayout.java
@@ -740,7 +740,7 @@ public FontMetrics getLineMetrics (int lineIndex) {
 			height = Math.max(height, ascent + descent);
 		}
 	}
-	return FontMetrics.wpf_new((int)baseline, (int)(height - baseline), 0, 0, (int)height);
+	return FontMetrics.wpf_new((int)baseline, (int)height - (int)baseline, 0, 0, (int)height);
 }
 
 /**
